Move the search classes into a more logical place.
