+++ Sources/StoryBB/Search/API_Interface.php
@@ -8,10 +8,9 @@
  * @version 3.0 Alpha 1
  */
 
-/**
- * Interface search_api_interface
- */
-interface search_api_interface
+namespace StoryBB\Search;
+
+interface API_Interface
 {
 	/**
 	 * Check whether the specific search operation can be performed by this API.
@@ -131,96 +130,3 @@ interface search_api_interface
 	 */
 	public function searchQuery(array $query_params, array $searchWords, array $excludedIndexWords, array &$participants, array &$searchArray);
 }
-
-/**
- * Class search_api
- */
-abstract class search_api implements search_api_interface
-{
-	/**
-	 * @var string The last version of SMF that this was tested on. Helps protect against API changes.
-	 */
-	public $version_compatible = 'StoryBB 3.0 Alpha 1';
-
-	/**
-	 * @var string The minimum SMF version that this will work with
-	 */
-	public $min_smf_version = 'StoryBB 3.0 Alpha 1';
-
-	/**
-	 * @var bool Whether or not it's supported
-	 */
-	public $is_supported = true;
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function isValid()
-	{
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function searchSort($a, $b)
-	{
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function prepareIndexes($word, array &$wordsSearch, array &$wordsExclude, $isExcluded)
-	{
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function indexedWordQuery(array $words, array $search_data)
-	{
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function postCreated(array &$msgOptions, array &$topicOptions, array &$posterOptions)
-	{
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function postModified(array &$msgOptions, array &$topicOptions, array &$posterOptions)
-	{
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function postRemoved($id_msg)
-	{
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function topicsRemoved(array $topics)
-	{
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function topicsMoved(array $topics, $board_to)
-	{
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function searchQuery(array $query_params, array $searchWords, array $excludedIndexWords, array &$participants, array &$searchArray)
-	{
-	}
-}
-
-?>
